Allowing the usage of input image of type float

+++ test/CoocurrenceTextureFeaturesImageFilterTestVectorImageSeparateFeatures.cxx
@@ -47,7 +47,7 @@ int CoocurrenceTextureFeaturesImageFilterTestVectorImageSeparateFeatures( int ar
   const unsigned int VectorComponentDimension = 8;
 
   // Declare types
-  typedef int   InputPixelType;
+  typedef float InputPixelType;
   typedef float OutputPixelType;
 
   typedef itk::Image< InputPixelType, ImageDimension >        InputImageType;
